Refactored the player box based on the stream status
